\name{ddp4hc-package}
\alias{ddp4hc-package}
\alias{ddp4hc}
\docType{package}
\title{
\packageTitle{ddp4hc}
}
\description{
\packageDescription{ddp4hc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ddp4hc}
\packageIndices{ddp4hc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ddp4hc}

Maintainer: \packageMaintainer{ddp4hc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
