% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DDPM_Linear_Ca}
\alias{DDPM_Linear_Ca}
\title{DDPM method with Linear regression model and Cauchy prior for concentration parameter for DP}
\usage{
DDPM_Linear_Ca(
  y_C,
  x_C,
  g_C,
  y_h,
  x_h,
  study_id_h,
  hcauchy_scale = 1,
  mu_beta_G0 = 0,
  tau_beta_G0 = 10000,
  alpha_sigma_G0 = 1,
  beta_sigma_G0 = 1,
  proposal_phi_sd = 0.1,
  proposal_M_DP_sd = 1,
  NBURN = 4000L,
  NTHIN = 10L,
  NOUTSAMPLE = 4000L,
  print = 0L
)
}
\arguments{
\item{y_C}{Response variable for Current trial}

\item{x_C}{covariate matrix for Current trial}

\item{g_C}{Group identifier Current trial}

\item{y_h}{Response variable for historical control}

\item{x_h}{covariate matrix for historical control}

\item{study_id_h}{study indicator for historical control}

\item{hcauchy_scale}{scale parameter for concentration parameter's prior in a DP, with a default value of 1}

\item{mu_beta_G0}{mean parameter for regression coefficient's prior, with a default value of 0}

\item{tau_beta_G0}{variance parameter for regression coefficient's prior, with a default value of 10000}

\item{alpha_sigma_G0}{shape parameter of Gamma prior for error variance with a default value of 1}

\item{beta_sigma_G0}{scale parameter of Gamma prior for error variance with a default value of 1}

\item{proposal_phi_sd}{standard deviation of truncated normal distribution as a proposal for phi, with a default value of 0.1}

\item{NBURN}{Number of iterations for the burn-in phase, with a default value of 4000}

\item{NTHIN}{Thinning rate applied to the chain, with a default value of 10}

\item{NOUTSAMPLE}{Mumber of iterations in the output chain, with a default value of 4000}

\item{print}{Specifies whether to display the MCMC output during execution with a default value of 0 (none)}
}
\description{
DDPM method with Linear regression model and Cauchy prior for concentration parameter for DP
}
\examples{
# No examples provided yet.

}
