% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DPM_Bin_Ca}
\alias{DPM_Bin_Ca}
\title{DPM method with binomial model for aggregated study-level data and Cauchy prior for concentration parameter for DP}
\usage{
DPM_Bin_Ca(
  y,
  n,
  hcauchy_scale = 1,
  proposal_M_DP_sd = 1,
  NBURN = 4000L,
  NTHIN = 10L,
  NOUTSAMPLE = 4000L,
  print = 0L
)
}
\arguments{
\item{y}{Response variable vector where the last element corresponds to current control data}

\item{n}{Number of participant vector where the last element corresponds to current control data}

\item{hcauchy_scale}{scale parameter for concentration parameter's prior in a DP, with a default value of 1}

\item{proposal_M_DP_sd}{proposal for M_DP, with a default value of 1}

\item{NBURN}{Number of iterations for the burn-in phase, with a default value of 4000}

\item{NTHIN}{Thinning rate applied to the chain, with a default value of 10}

\item{NOUTSAMPLE}{Mumber of iterations in the output chain, with a default value of 4000}

\item{print}{Specifies whether to display the MCMC output during execution with a default value of 0 (none)}
}
\description{
DPM method with binomial model for aggregated study-level data and Cauchy prior for concentration parameter for DP
}
\examples{
# No examples provided yet.

}
