% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DPM_Cont}
\alias{DPM_Cont}
\title{DPM method with normal model for aggregated study-level continuous data}
\usage{
DPM_Cont(
  ybar,
  sd,
  n,
  hyper_gamma_shape = 1,
  hyper_gamma_scale = 1,
  mu_G0 = 0,
  tau2_G0 = 10000,
  NBURN = 4000L,
  NTHIN = 10L,
  NOUTSAMPLE = 4000L,
  print = 0L
)
}
\arguments{
\item{ybar}{Sample mean vector where the last element corresponds to current control data}

\item{sd}{Sample standard deviation vector where the last element corresponds to current control data}

\item{n}{Number of participant vector where the last element corresponds to current control data}

\item{hyper_gamma_shape}{Shape parameter for concentration parameter's prior in a DP, with a default value of 1}

\item{hyper_gamma_scale}{Scale parameter for concentration parameter's prior in a DP, with a default value of 1}

\item{mu_G0}{Mean parameter for component mean's prior, with a default value of 0}

\item{tau2_G0}{Variance parameter for component mean's prior, with a default value of 10000}

\item{NBURN}{Number of iterations for the burn-in phase, with a default value of 4000}

\item{NTHIN}{Thinning rate applied to the chain, with a default value of 10}

\item{NOUTSAMPLE}{Number of iterations in the output chain, with a default value of 4000}

\item{print}{Specifies whether to display the MCMC output during execution with a default value of 0 (none)}
}
\description{
DPM method with normal model for aggregated study-level continuous data
}
\examples{
ybar <- c(8.7, 8.0, 4.4, 6.1, 5.1, 4.8)
sd   <- c(7.2, 5.8, 6.4, 7.1, 6.8, 6.3)
n    <- c(111, 202, 169, 63, 164, 55)

fit <- DPM_Cont(
  ybar = ybar,
  sd = sd,
  n = n,
  NBURN = 100,
  NTHIN = 2,
  NOUTSAMPLE = 100,
  print = 0
)

}
