% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Simple_Linear}
\alias{Simple_Linear}
\title{Linear regression model}
\usage{
Simple_Linear(
  y,
  x,
  g,
  mu_beta_G0 = 0,
  tau_beta_G0 = 10000,
  alpha_sigma_G0 = 1,
  beta_sigma_G0 = 1,
  NBURN = 4000L,
  NTHIN = 10L,
  NOUTSAMPLE = 4000L,
  print = 0L
)
}
\arguments{
\item{y}{Response variable}

\item{x}{Predictor matrix}

\item{g}{Group identifier}

\item{mu_beta_G0}{mean parameter for regression coefficient's prior, with a default value of 0}

\item{tau_beta_G0}{variance parameter for regression coefficient's prior, with a default value of 10000}

\item{alpha_sigma_G0}{shape parameter of Gamma prior for error variance with a default value of 1}

\item{beta_sigma_G0}{scale parameter of Gamma prior for error variance with a default value of 1}

\item{NBURN}{Number of iterations for the burn-in phase, with a default value of 4000}

\item{NTHIN}{Thinning rate applied to the chain, with a default value of 10}

\item{NOUTSAMPLE}{Mumber of iterations in the output chain, with a default value of 4000}

\item{print}{Specifies whether to display the MCMC output during execution with a default value of 0 (none)}
}
\description{
Linear regression model
}
\examples{
# No examples provided yet.

}
