\name{ddp4hc-package}
\alias{ddp4hc-package}
\alias{ddp4hc}
\docType{package}
\title{
\packageTitle{ddp4hc}
}
\description{
\packageDescription{ddp4hc}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{ddp4hc}

Maintainer: \packageMaintainer{ddp4hc}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = ddp4hc)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{ddp4hc}}
%\section{Documentation Index}{\packageIndices{ddp4hc}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
